fix package names. Move sharpness to library
